requirements.txt
Django>=3.1.6,<3.2.0
djangorestframework>=3.12.2,<3.13.0
djangorestframework-jwt>=1.11.0,<1.12.0
dj-config-url==0.1.1
psycopg2>=2.8.6,<2.9.0

